# docs/swagger.yaml
definitions:
...
  api.ListUserRequest:
    properties:
      page_id:
        minimum: 1
        type: integer
      page_size:
        maximum: 10
        minimum: 5
        type: integer
    required:
    - page_id
    - page_size
    type: object
  api.ListUserStudentRequest:
    properties:
      page_id:
        minimum: 1
        type: integer
      page_size:
        maximum: 10
        minimum: 5
        type: integer
    required:
    - page_id
    - page_size
    type: object
  api.ListUserTeacherRequest:
    properties:
      page_id:
        minimum: 1
        type: integer
      page_size:
        maximum: 10
        minimum: 5
        type: integer
    required:
    - page_id
    - page_size
    type: object
...
  api.UpdateUserRequest:
    type: object
...
  api.createUserRequest:
    properties:
      email:
        type: string
      full_name:
        type: string
      hashed_password:
        type: string
      qualification:
        type: string
      role:
        type: string
      user_name:
        type: string
    type: object
  api.loginUserRequest:
    properties:
      hashed_password:
        type: string
      user_name:
        type: string
    type: object
...
info:
  contact: {}
paths:
  /admin/signup:
    post:
      consumes:
      - application/json
      description: Create a new admin's user with the provided details
      operationId: create admin user
      parameters:
      - description: User creation request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/api.createUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      summary: Create a admin's user
...
  /listadmin:
    get:
      consumes:
      - application/json
      description: ListUser with the provided admin based
      operationId: list-user
      parameters:
      - description: admin list request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/api.ListUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      summary: ListUser
  /liststudent:
    get:
      consumes:
      - application/json
      description: ListUserStudent with the provided student based
      operationId: list-student
      parameters:
      - description: student list request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/api.ListUserStudentRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      summary: ListUserStudent
  /listteacher:
    get:
      consumes:
      - application/json
      description: ListUserTeacher with the provided teacher based
      operationId: list-teacher
      parameters:
      - description: teacher list request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/api.ListUserTeacherRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      summary: ListUserTeacher
  /login:
    post:
      consumes:
      - application/json
      description: Log in a user with the provided credentials
      operationId: login-user
      parameters:
      - description: Login request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/api.loginUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      summary: Log in user
...
  /signup:
    post:
      consumes:
      - application/json
      description: Create a new user with the provided details
      operationId: create-user
      parameters:
      - description: User creation request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/api.createUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      summary: Create a new user
  /submissions:
    get:
      description: Lists submissions for a given assignment ID
      operationId: listSubmissions
      parameters:
      - description: Assignment ID
        in: path
        name: assignment_id
        required: true
        type: integer
      - description: Limit
        in: query
        maximum: 100
        minimum: 1
        name: limit
        required: true
        type: integer
      - description: Offset
        in: query
        minimum: 0
        name: offset
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      summary: List submissions
  /submissions/{assignment_id}/{student_id}:
    get:
      description: Retrieves a submission by assignment and student ID
      operationId: getSubmission
      parameters:
      - description: Assignment ID
        in: path
        name: assignment_id
        required: true
        type: integer
      - description: Student ID
        in: path
        name: student_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      summary: Get a submission
...
  /user/edit:
    patch:
      consumes:
      - application/json
      description: Updates a user with provided details
      parameters:
      - description: Updated user details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/api.UpdateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      summary: Update a user
swagger: "2.0"